+++ KmpIO/src/linuxTest/kotlin/test/ZipFileBasics.kt
@@ -52,4 +52,9 @@ class ZipFileBasics {
         tests.zipDirectoryTest(false)
         tests.zipDirectoryTest(true)
     }
+
+    @Test
+    fun ntfsTagTest() {
+        tests.ntfsTagTest()
+    }
 }